cdss_django/requirements.txt
acme==1.21.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==21.2.0
Babel==2.8.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.4
cachetools==5.5.2
celery==5.5.3
certbot==1.21.0
certbot-apache==1.21.0
certifi==2025.4.26
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==25.1.2
cloudpickle==3.1.1
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.5.1
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kombu==5.5.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.44.0
MarkupSafe==2.0.1
more-itertools==8.10.0
mysqlclient==2.2.7
netifaces==0.11.0
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.0
packaging==21.3
pandas==2.3.0
parsedatetime==2.6
pexpect==4.8.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydicom==3.0.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-augeas==0.5.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.32.4
requests-toolbelt==0.9.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
SecretStorage==3.3.1
shap==0.47.2
six==1.16.0
slicer==0.0.8
sos==4.8.2
soupsieve==2.7
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
websocket-client==1.2.3
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
